Resolves #34 and additions to #17
The final level sequence issue is now resolved (#34), so for example the 4th level in TR2G will always be the last level, and the 5th can only be played if all secrets are found.
Lots of work on the Audio functionality, mainly extracting existing audio from TR2/3 and creating smaller WAD files for this system to reference. The data files are stored separately but point to the WAD files to allow sampling.