#47, #48 Compression and sunsets
The config files are now compressed (#48). The sunset flag can now be set for any level, and this also changes the room lighting to 3 for each room in that level where it's set to 0. Testing to be carried out yet i.e. leaving a level running for 20 minutes to check that the lighting fades.