#65 Lock Randomisation
Global randomisation supported flag is now in place, meaning that callers can decide whether or not to support it. If it's not supported, the idea is that any previous randomisations performed by other extensions will remain in place, so the caller that doesn't support it should ensure that any options to organise are hidden.