#66 Save cancellation
Saves can now be cancelled. All writes are now performed to a temporary WIP directory and only on full save transaction success will the output files be moved to the target directory. Callers can cancel the save progress at any stage, and so the integrity of existing files is maintained.