#87 Read-only file backup workaround
Puts in place a workaround if a read-only file has been copied to the backup folder by ensuring that all files that do arrive in the backup folder have this flag removed. This means if a data directory contains RO files, the system will no longer throw an exception when it tries to clear out the files it doesn't need.
This is a workaround until the whole backup method is improved i.e. only the required files should be copied in the first place.