#98 TR3 Support
Separates TR23ScriptedLevel into TR2ScriptedLevel and TR3ScriptedLevel. This is a result of secret counts being hardcoded in the game for TR3 and so as a result we need to differentiate NumSecrets between TR2 and TR3. TR2 is also hardcoded, but it can be either 3 or 0, whereas TR3 is based on the level sequencing.